refactor gtp.erl, split into socket, apn and dp modules
The main modules gtp.erl was doing to much in one place and
mixing the data path, control path and APN functions is limiting
the implementation.

In a future version, the control path should be able to handle
multiple data path forwarding elements. Multi APN requires
per APN settings and IP pools.

To prepare for all that, split gtp.erl into modules and API's
for DP, APN and control sockets.