rewrite handling of recovery elements
Recovery element handling is unified accross all interface
implementation

Incomming recovery elements are now reported to the path
and the path restart counter is recorded in the local
context. Currently no action is taken when a peer restart is
detected (that will come in a later step).

If we assume that the remote peer does not yet have our
restart counter a recovery element will be attached to
eligible messages.

A remote peer is assumed to need our restart counter when
either:
* no restart counter for that peer is know
* the peers send us a recovery element