rework proxy map interface to take defaults from handler
Instead of testing wether a five value was modified in the ds map
handler, preload the proxy info with the values from request.
This simplyfies the API and allows us to pass other values through.

Use that to pass the default GGSN IP and allow the ds map procedure
to change it.