process enviroment configuration through setup
The setup application is already pulled in through exometer_core.
Without configuration, setup will create data and log directories
in the application home directory. In order to keep a installed
ergw directory clean, the setup application need to be configured
with sensible log and data paths.
Since we already need to handle configuration for the setup app,
we can use some of its feature as well.
This change applies setup's variable expansion to the state_file
and the other configuration settings (see CONFIG.md for deatils).