add reuseaddr socket option and properly close GTP sockets
GTP sockets need to be properly closed on shutdown to be able
to reuse them. This add the reuseaddr socket option to avoid
having to wait for some time after socket close to reopen the
same port and it also makes sure we close the inet socket when
the gtp_socket process is terminated.