test: add basic test suite for PGW proxy
The proxy test connects to an erGW based PGW. To allow to
run the proxy and the PGW on the same localhost IP, two
hidden configuration options have been added to run GTP
sockets on non-default UDP ports.

Also, the gtp_context terminate function has been extended
to the context protocol handlers to be able to syncronize
the test to the context shutdown and validate the successful
shutdown.