fix path restart behavior
A message with a restart counter that causes a reset of a
path needs to be registered *AFTER* all existing PDP contexts
have been cleared.

Also make sure that we don't mix up the old and new pid. The
ets traversal run in a new process, swap ets tables to make
sure that this process is not seeing newly added pid's.

Echo Requests are handeled by the GTP path. It therefore also
needs to handle the recovery counters in them.