test: rework the test suite to properly handle the restart_counter
A change of the restart_counter will cause the path to remove all
PDP context. This means we need to be careful about seperating the
path and to send sensible restart_counter.

Also test that a restart_counter change kills existing contexts,
and that request message that contain a counter change are handled
properly.