handle resents for outstanding requests properly
In proxy mode, re-sends should not reprocess the original
request but trigger a retransmission of the forwarded
request.
That logic was only partially implemented and what was there
was mostly broken.

Cache send requests that have a request id and use the cached
request for when a resend is triggered.

Add tests to ensure the resend logic is working in proxy mode.