ensure all pending requests are removed from the request table
We register the pending requests in the context registry. Unlike,
TEID's, IMSI's and so, these are transient events and need to be
remove once a request is finished. It doesn't mater if the request
was abort or completed, it was to be removed when processing it is
done.

Command have to clean up immediately, requests should wait for a
response to be finished.