change Modify Bearer Request SGW change detection for RAU/TAU/HO
Instead of checking for an actual IP or TEID change, take the pure
presens of the FQ-TEID element as SGW change indication.

3GPP TS 29.274, Sect. 7.2.7 Modify Bearer Request says that we should
consider the content as well, but in practice that is not stable enough
in the presense of middle boxes between the SGW and the PGW

Add test to validate that all required answer element are present on
SGW change.