fix racy termination of colliding session
It was possible that the termination and deregistration of a
colliding session had not finished when we attempted to register
the new session. This resulted in a 'System Error' response.

Change the contrex registry and add await_dereg function to ensure
the old session has been properly removed before attempting a new
registration.

Add a test case.