add primary key to context and use it to detect collisions
A GTP context can only have one primary key, either with the IMSI
or the IMEI. The NSAPI/LBI has to present on context establishment,
but some follow-up requests need to find the context without the
NSAPI/LBI.

Contruct the primary key from IMSI or IMEI and NSAPI/EBI. Find it
by using a match that also works when the NSAPI/LBI is unknown.

Adjust to test to the new key and fix the context collisions in them.