let the protocol handler specify the destionation port for a request
GTPv2 triggered requests need to be send to the port from where the
trigger command originated. Add a port option to the send_requests
methods and use the default port for now. This move the default port
from gtp_socket into the handlers.

Setting the correct destination port for a triggered requests will
come in a future commit.