Initial version of redirector mode for GTP sockets
Now it is possible to create GTP socket in a redirection mode.

The redirection mode means that a GTP request will be transferred to
a GTP backed with saving source IP and port. For the GTP backend such
request looks like it came from particular GTP client directly.

The following features are supported:

* Node memorization to guarantee that retransmitted request goes
  to previously selected processing node
* Add GTP-Redirector counter metric. The `rr` metric counts the number
  of message processed in GTP redirector mode.
* Keep-Alive mechanism for GTP-Redirector.

Note that IPv6 is not implemented yet.