rework S11 GTP-U TEID handling and Sx node selection
For initial attachement a Create Session Request on S11 does not
contian the GTP-U TEID. The GTP-U TEID will be set with follow up
Modify Session Request. This is very similar how a SGW or proxy
session starts.
Adjust the SEA logic to match the SGW processes, simplify session
creation in the process, fix test suite to test the Create/Modify
sequence and sync Sx node selection with the other implementations.