fix IPv6 prefix and Interface Identifier handling in IPv6 bearers
For a IPv6 bearer we need to assign a /64 prefix and a interface
indentifier. The interface can be any value as long as it does
not conflict our local interface id and it can't be a reserved
value.

Turns out that zero is a reserved interface id and we where using
that. Use 1 instead.

Also strenghten the pool validation.

Experimatal support to permit other prefixes than /64 to be used.
Doing so is no going to work on 3GPP EPC, however it might work
on other media.