DNS resolver: fix issue when after NAPTR we get A record with IP
Despite [RFC 3958](https://tools.ietf.org/html/rfc3958#section-6.4)
says:

```
"A" means that the output of the Rule is a domain name and should
be used to lookup address records for that domain.
```

Result of NAPTR may be A record with IP address as a replacement.
In that case we just need to use that IP as result of lookup procedure.