fix cache expire logic in GTP-C and Sx sockets
Calling ergw_cache:expire/1 outside of an timeout event will
start a new timer regardless of wether there is already one
running.
Also, there is really no need to invoke expire outside of an
timer event. The normal internal cleanup in enter/4 should be
sufficient and the timer events will handle the rest.

Change the expire API so that it can deals with the timers
properly.