unify context deletion on PFCP during session teardown
Issuing a PFCP Deletion Request and handling the returned
usage report is the same for all protocol handlers. Move into
a central place and use it anywhere.

Add a reason to skip the actual context deletion in cases when
the UPF has already removed the context (e.g. UPF restart).