fix handling of GTP Echo Requests in gtp_path
A Echo Request timeout will NOT return a #gtp{} message, therefor
matching on that in handle_response is wrong.

Add a additional parameter to the response callback to make sure
we only handle echos and let the echo_response logic deal with
the timeout (as it was supposed to in the first place).