support fully qualified domain names in APN selection lookups
The using the NAPTR based node selection, only APNs and 3GPP domains
where allowed. Those where subject to nomalization that only works
on this type of names.
To allow abritray node selection in private setups, a special taged
FQDN is now allowed instead off APNs and forward addresses. This is
intended mostly for the proxy handlers.