handle node_selection option in proxy contexts
Change the rules for node selection on proxy contexts and
handle the option for those context properly.

Before this change the node selection on proxy sockets was
not supported, but the code was expecting one. This resulted
in an empty selection list.

Now, proxy context can have either a proxy socket, a node
selection or both options set.