implement SGSN/SGW change forwarding in proxy mode
In order to make a SGSN/SGW change visible through a proxy,
the proxy must also change its control TEID on the forwarded
connection.

It turns out that this *MUCH* simple that initialy thought,
because a) only Update PDP Context and Modify Session requests
are actually used in SGSN/SGW change and b) only the control
TEIDs need to be changed. The data TEIDs can be left alone.