add proper PGW/GGSN policy config to proxy tests
The GGSN/PGW config for the destination GSN is taken from the main
test config. We therefor need a valid policy config in there or
the test will fail.
No having the config in there will not cause a problem when all test
suites are run since the policy config from the normal PGW/GGSN cases
will leak into the proxy tests. But running only the proxy suites
will fail.