Fix metric name of IP pool in metrics
8cf5938 introduced IP pool metrics. Upon accessing these metrics via the
Prometheus format, the following invalid metric name is created:

    pool_<<"sgi">>_ipv4_10_0_0_1_free 262112

The invalid part `<<"sgi">>` is caused by http_api_handler:ioize/1. This
commit fixes the function to produce:

    pool_sgi_ipv4_10_0_0_1_free 262112

In addition, pools for IPv6 addresses would create an invalid metric
name for Prometheus (see [1]). By converting the ":" to "_", this is
fixed as well.

[1]: https://prometheus.io/docs/concepts/data_model/#metric-names-and-labels