major rework of the PCC rules handling
* move rules and charging information into dedicated context
  (instead of keeping it in the AAA session)
* calculate PFCP change events for URRs based on differences in
  old vs. new context

This primarily reworks how the online charging releated charging
events in Gy are handled. The old implmentation was focused on
events triggered by a Gy interaction. Events that resulted from
other API (e.g. adding/removeing a rating group as result of
a Gx interaction) where not properly handled and hard to integrate
with the old implementaion.

This new version keep track a of the current state off policy and
charging information and generates the PFCP update information based
on the differences.