rework APN NI/OI use in proxy
When selecting a forwarding target the APN-OI has taken to be in
consideration. For GTPv2 the OI is required in APN IE, for GTPv1
the OI is optional (TS 29.060 isn't very clear on that, but in
the field the OI is not part of Create PDP Requests).
If no OI is given, calculate it from the IMSI afte applying the
proxy info.