handle subtly broken GTP message in gtp_socket
A GSN that messes up the flags in GTP v1 Echo Requests was observed
in wild. The broken flags turn that message in a GTP' message.
With an older gtplib that message would fail decoding and be
discarded. Newer version do decode GTP' and would lead to crash when
attempting to handle the unsupported version.