make the IPv6 UE interface id configurable
Using a hard coded IPv6 interface id of 1 for the UE is not always
desirable. Make it configurable, choices are: default (1), random,
a give value or sgsnemu.

The sgsnemu option uses the 64 bit IPv6 prefix as the interface id.
This is concession to the way that sgsnemu handles it IPv6 contest
lookup and should allow IPv6 only bearers with sgsnemu.