add tini PID 1 to ercdf container
Usually, a process spawning subprocesses is responsible for cleaning
them up once they finish. This responsibility propagates down to PID 1.

When finished subprocesses ("childs") are not reaped, they become zombies
which are never vanish. Over time, they can starve the system. Erlang and
Elixir does not take care of spawned children on its own. Thus, any
"remote console" interaction with a running Erlang node will leave behind
a couple of zombies when PID 1 is not taking care of them.
Inside the container, Erlang / Elixir is usually run as PID 1.

This commit adds tini as the new PID 1 of the containers. As a result, all
child processes will be reaped properly.