implement path monitoring profiles for outgoing (active) paths
For an active path, the echo request interval is adjusted depending on
wether the path as active contexts (busy), no contexts (idle) or is
down.
For idle and down paths a maximum time to keep their state, monitor
them or try to reconntact them is introduced.

A path is considered active when we have initiated it (e.g. SGSN/SGW or
proxy use).