rework traking of active GTP paths
Only GTP paths with active contexts should be actively path monitored.

A context is considered to be active when we have received a request
for it or if a peer has responded to request for this context from us.
A context is not (yet) active while we are trying to send an initial
request to a remote peer.