simplify tei mngr, drop ets from it
Entries from the TEI table need to removed, but are not.

tei mngr was imported from the stateless branch. That brought with
it assumtions about how session state is keept and removed which
do not apply to the current ergw state.
Attempting to make them fit, results in large, structural changes
that are not really suitable for this ergw version.

Fix the ETS leak by removing the table and use the context registry
to check for existing TEIDs.