replace local_data_endp in context with left#bearer.local
The local_data_endp member is used to record the local data GTP
session on the left side of the GSN. Replace this with a more
generic entry in a bearer structure.

The bearer structure will record the local and remote side of
the bearer (tunnel) in a generic way.