fix reading from socket error queue
A select info is generated whenever there is something to read from
the socket, in C terms `(POLLIN | POLLERR)`. There is no distinction
whether that is normal read or errqueue read (POLLERR).
This adds a heuristic to read the errqueue should we get no data from
a normal socket read.