CT: change the nested groups in the GTP suites
This makes the group hierarchy on the non proxy cases match the
proxy tests. It has ipv4 and ipv6 as root groups and nested
groups underneath.

Especially the PGW suite benefits from that. Selecting to run the
common IPv4 test suite was not successfull with the old ordering.
With the new setup `rebar3 ct .. --group=ipv4,common` works.