fix statem data corruption in S5/S8 proxy
The corruption of the statem data only happens when trying to close
the context. This was the cause for many of the CI failures in
pgw_proxy_SUITE.dns_node_selection.

While this might ocure in running instance as well, there should be
not operational impact as it only abort a context that is already
terminating.