add a `suspect` state to gtp peers
The new `suspect` state has it's own echo request interim settings
and a timeout. The timeouts in `idle` and `down` trigger a path stop
action, the timeout in `suspect` state tiggers a transition into
`down` with all the actions that are assoziated with going to down
(terminate all gtp contexts).

A echo timeout or ICMP error will no longer transition directly into
the `down` state.

A GTP path in state suspect will not drop contexts, the state timeout
and the following transition into `down? will do that.

A timeout of 0 (zero) for `suspect` will result in an immediate transition
into `down`. This mirrors the behavior before this change.