refactor app layout and startup
Application startup is move out of ergw_core into a new (old) ergw
application.

ergw_core is now only a library that needs to be initialized properly
from the main app.
ergw is responsible for reading and ypplying the configuration to
ergw_core.

This is the first step of untangling the root config from ergw_core. The
test suites are currently broken and will be converted in the next steps.