adjust cache expiry interval
The timeout argument in ergw_cache:new/2 define the interval at
which expired entries are pruned from the cache. Setting to high
values means that potentially many entries need to be removed
from the cache at once. Lower values mean that expiration check
is run more often and as to deal with less entries in each run.

Processing a few entries at a time is generally preferable to
handle many at once.

Change the expiry interval of all users to 100ms.