replace map in IP pool registry with ets
The map required a synchronous call to the server. The ets table
allows for concurent reads without synchnoizing the reader processes.

The benchmark suite shows no speed up from. However that is not
surprising as the results from the suite do vary heavily between
runs.

The registry showever up in a observer output with a non-zero
message-q during benchmark runs. This change makes that go away.