rewrite GTP-v2 CSR to non-blocking FSM
Use a hand writen function chaining to replace the deep function
returns with error returns.
Use the same chaining mechanism to make it possible to halt and
restart a procedure in the midle of execution. This is then used to
wait for the results of blocking calls in a non-blockin fashion.