Suppress reporting of ergw_cluster startup process termination
At startup of ergw_cluster process it creates a special process to
be sure that Erlang runtime fully started before ergw_cluster may
start to do its job. After that the special process done with its
job we terminate it with exit(ok) and this event is reported to log.

According to proc_lib documentation:

> When a process that is started using proc_lib terminates abnormally
  (that is, with another exit reason than normal, shutdown, or {shutdown,Term}),
  a crash report is generated, which is written to terminal by the default
  logger handler setup by Kernel

This commit replaces proc_lib:spawn_link/1 with erlang:spawn_link/1 to get rid
of this report in console. There is no sense to have this report in logs as
anyway the exit signal will be handled by the ergw_cluster process.