add temp. workarround for out of sequence PFCP usage reports
This a a know problem with the sequencing of GTP location updates that
arrive simultaneously with a context teardown action.
The location update triggeres a PFCP Modify Session, the teardown a
PFCP Delete Session, the Modify Session Response can then arrive after
the teardown action has already removed the PFCP context reference.