Do not add a monitor if a PID is either in the contexts or monitors map.
Currently, a PID is not monitored if it is present in both the contexts
and monitors maps. However, there are cases where a monitor is created
in the monitors list in a register call and then moved to the contexts
list in a bind call, so a new monitor on the same process is created but
not tracked. After a time these monitors fill up the memory and crash the
node.

This commit will not add a new monitor if the PID is present either in
the contexts or the monitors maps.